Split tests of input.py into multiple files
